# [apps][templates] integrate core autolinking from our implementation