--- a/apps/bare-expo/android/settings.gradle
+++ b/apps/bare-expo/android/settings.gradle
@@ -2,8 +2,26 @@ pluginManagement {
   includeBuild(new File(["node", "--print", "require.resolve('@react-native/gradle-plugin/package.json')"].execute(null, rootDir).text.trim()).getParentFile().toString())
 }
 plugins { id("com.facebook.react.settings") }
-extensions.configure(com.facebook.react.ReactSettingsExtension){ ex ->
-        ex.autolinkLibrariesFromCommand()
+
+def useUnstableCoreAutolinking = System.getenv('EXPO_UNSTABLE_CORE_AUTOLINKING')
+useUnstableCoreAutolinking = '1'
+
+extensions.configure(com.facebook.react.ReactSettingsExtension) { ex ->
+  if (useUnstableCoreAutolinking == '1') {
+    def command = [
+      'node',
+      '--no-warnings',
+      '--eval',
+      'require(require.resolve(\'expo-modules-autolinking\', { paths: [require.resolve(\'expo/package.json\')] }))(process.argv.slice(1))',
+      'react-native-config',
+      '--json',
+      '--platform',
+      'android'
+    ].toList()
+    ex.autolinkLibrariesFromCommand(command)
+  } else {
+    ex.autolinkLibrariesFromCommand()
+  }
 }
 
 rootProject.name = 'BareExpo'