--- a/apps/fabric-tester/ios/Podfile
+++ b/apps/fabric-tester/ios/Podfile
@@ -18,6 +18,7 @@ target 'fabrictester' do
   use_expo_modules!
 
   if ENV['EXPO_UNSTABLE_CORE_AUTOLINKING'] == '1'
+    Pod::UI.puts('Using expo-modules-autolinking as core autolinking source'.green)
     config_command = [
       'node',
       '--no-warnings',