--- a/apps/native-tests/ios/Podfile
+++ b/apps/native-tests/ios/Podfile
@@ -4,6 +4,8 @@ require File.join(File.dirname(`node --print "require.resolve('react-native/pack
 require 'json'
 podfile_properties = JSON.parse(File.read('./Podfile.properties.json')) rescue {}
 
+ENV['EXPO_UNSTABLE_CORE_AUTOLINKING'] = '1'
+
 prepare_react_native_project!
 
 platform :ios, podfile_properties['ios.deploymentTarget'] || '15.1'
@@ -20,7 +22,21 @@ target 'NativeTests' do
 
   pod 'ExpoModulesTestCore', :path => "../../../packages/expo-modules-test-core/ios"
 
-  config = use_native_modules!
+  if ENV['EXPO_UNSTABLE_CORE_AUTOLINKING'] == '1'
+    config_command = [
+      'node',
+      '--no-warnings',
+      '--eval',
+      'require(require.resolve(\'expo-modules-autolinking\', { paths: [require.resolve(\'expo/package.json\')] }))(process.argv.slice(1))',
+      'react-native-config',
+      '--json',
+      '--platform',
+      'ios'
+    ]
+    config = use_native_modules!(config_command)
+  else
+    config = use_native_modules!
+  end
 
   use_frameworks! :linkage => podfile_properties['ios.useFrameworks'].to_sym if podfile_properties['ios.useFrameworks']