… apps/bare-expo/ios/Podfile.lock
… apps/fabric-tester/ios/Podfile.lock
… apps/native-tests/ios/Podfile.lock