# Why

integrate with our autolinking for react-native core

# How

- use custom autolinking command from react-native core
- add `EXPO_UNSTABLE_CORE_AUTOLINKING` environment variable as a feature flag
- enable `EXPO_UNSTABLE_CORE_AUTOLINKING` on bare-expo and fabric-tester. the expo-template-bare-minimum is default off still.

# Test Plan

- check bare-expo Podfile.lock before and after `EXPO_UNSTABLE_CORE_AUTOLINKING`
- bare-expo `./gradlew :app:dependencies | sort > dep.log`, diff the dep.log before and after `EXPO_UNSTABLE_CORE_AUTOLINKING`
- benchmark
  - bare-expo ios `pod install`: 15.53s -> 14.21s
  - bare-expo android `./gradlew :app:assembleDebug`: 16.40s -> 14.27s  

# Checklist

- n/a Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [x] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).